refactor: 각 컴포넌트의 reducer파일 위치이동
- reducer는 state에 의존적이므로 state가 위치한 곳에 있어야한다고 판단해서 파일 위치를 수정함.
