+++ src/components/ImageEditor/index.js
@@ -1,4 +1,3 @@
 export { default } from './ImageEditor';
 import * as action from './action';
-import * as reducer from './reducer';
-export { action, reducer };
+export { action };
