+++ src/components/PostUploader/ImageUploader/index.js
@@ -1,4 +1,3 @@
 export { default } from './ImageUploader';
-import * as reducer from './reducer';
 import * as action from './action';
-export { action, reducer };
+export { action };
